refactor(frontend): decouple source modal types
