+++ apps/frontend/src/components/sourcesModalTypes.ts
@@ -1,0 +1,6 @@
+export type SourceFilter = "all" | "enabled" | "disabled" | "attention";
+
+export interface Feedback {
+  tone: "error" | "success" | "warning";
+  message: string;
+}
